+++ R/rename_at.R
@@ -30,7 +30,7 @@
 #' @examples
 #' library(dplyr)
 #' recipe(~., data = iris) %>%
-#'   step_rename_at(everything(), fn = ~ gsub(".", "_", ., fixed = TRUE)) %>%
+#'   step_rename_at(all_predictors(), fn = ~ gsub(".", "_", ., fixed = TRUE)) %>%
 #'   prep() %>%
 #'   bake(new_data = NULL) %>%
 #'   slice(1:10)