+++ tests/testthat/test-colcheck.R
@@ -5,8 +5,8 @@ rp1 <- recipe(mtcars, cyl ~ .)
 rp2 <- recipe(mtcars, cyl ~ mpg + drat)
 
 test_that("check_col works in the prep stage", {
-  expect_error(rp1 %>% check_cols(everything()) %>% prep(), NA)
-  expect_error(rp2 %>% check_cols(everything()) %>% prep(), NA)
+  expect_error(rp1 %>% check_cols(all_predictors()) %>% prep(), NA)
+  expect_error(rp2 %>% check_cols(all_predictors()) %>% prep(), NA)
   expect_error(rp2 %>% check_cols(cyl, mpg, drat) %>% prep(), NA)
   expect_error(rp2 %>% check_cols(cyl, mpg) %>% prep(), NA)
 })
@@ -14,20 +14,20 @@ test_that("check_col works in the prep stage", {
 
 test_that("check_col works in the bake stage", {
 
-  expect_error(rp1 %>% check_cols(everything()) %>% prep() %>% bake(mtcars),
+  expect_error(rp1 %>% check_cols(all_predictors()) %>% prep() %>% bake(mtcars),
                NA)
-  expect_equal(rp1 %>% check_cols(everything()) %>% prep() %>% bake(mtcars),
+  expect_equal(rp1 %>% check_cols(all_predictors()) %>% prep() %>% bake(mtcars),
                tibble(mtcars[ ,c(1, 3:11, 2)]))
   expect_error(rp2 %>% check_cols(cyl, mpg, drat) %>% prep %>% bake(mtcars), NA)
   expect_equal(rp2 %>% check_cols(cyl, mpg, drat) %>% prep %>% bake(mtcars),
                tibble(mtcars[ ,c(1, 5, 2)]))
 
   expect_error(
-    rp1 %>% check_cols(everything()) %>% prep() %>% bake(mtcars),
+    rp1 %>% check_cols(all_predictors()) %>% prep() %>% bake(mtcars),
     NA
   )
   expect_equal(
-    rp1 %>% check_cols(everything()) %>% prep() %>% bake(mtcars),
+    rp1 %>% check_cols(all_predictors()) %>% prep() %>% bake(mtcars),
     tibble(mtcars[, c(1, 3:11, 2)])
   )
   expect_error(rp2 %>% check_cols(cyl, mpg, drat) %>% prep() %>% bake(mtcars), NA)
@@ -36,7 +36,7 @@ test_that("check_col works in the bake stage", {
     tibble(mtcars[, c(1, 5, 2)])
   )
   expect_snapshot(error = TRUE,
-    rp1 %>% check_cols(everything()) %>% prep() %>% bake(mtcars[-1])
+    rp1 %>% check_cols(all_predictors()) %>% prep() %>% bake(mtcars[-1])
   )
   expect_snapshot(error = TRUE,
     rp2 %>% check_cols(cyl, mpg, drat) %>% prep() %>%
@@ -195,7 +195,7 @@ test_that("empty selection tidy method works", {
 
 test_that("printing", {
   rec <- recipe(mpg ~ ., mtcars) %>%
-    check_cols(everything())
+    check_cols(all_predictors())
 
   expect_snapshot(print(rec))
   expect_snapshot(prep(rec))